Support Linux machines as well This adds a new library to the repo for use by linux machines. The polyblip.py script also has been changed to support two libraries.
